Refatorado(Task): dividido task.py em pipeline e cross-post
- Criado task_cross_post.py com todo o subsistema de publicação
  (executor, semáforo de fila, registro de Future, recuperação na
  inicialização, chamadas ao YouTube e Upload-Post)
- Reduzido task.py de 1620 para 1087 linhas, mantido como fachada
  e núcleo do pipeline (roteiro, áudio, legenda, material, vídeo
  final), reexportando o cross-post pra manter task.xxx() igual
- Corrigidos 5 patches de teste que trocavam nome de módulo inteiro
  (_cross_post_slots, _is_windows_process_alive, _run_cross_post)
  pra apontar pro módulo novo, já que a função consumidora migrou
  junto e passou a resolver o nome lá
